requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
tensorboard>=2.14.0
streamlit>=1.28.0
scikit-image>=0.21.0
pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0
kaggle>=1.5.0
faiss-cpu>=1.7.4
grad-cam>=1.5.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
opencv-python>=4.8.0
